[terminal] - login loading certificate after generation on creating user

+++ core/configuration/src/main/java/com/dcall/core/configuration/app/service/environ/EnvironServiceImpl.java
@@ -12,6 +12,7 @@ import com.dcall.core.configuration.generic.entity.cipher.CipherAES;
 import com.dcall.core.configuration.generic.entity.environ.Environ;
 import com.dcall.core.configuration.generic.entity.environ.EnvironBean;
 import com.dcall.core.configuration.generic.entity.hash.UserHash;
+import com.dcall.core.configuration.generic.entity.identity.Identity;
 import com.dcall.core.configuration.utils.FileUtils;
 import com.dcall.core.configuration.utils.ResourceUtils;
 import org.slf4j.Logger;
@@ -58,37 +59,39 @@ public class EnvironServiceImpl implements EnvironService {
             FileUtils.getInstance().createDirectory(getConfigDirectory());
             initHash(context);
         }
-
         final UserHash<String> userHash = context.getUserHash();
         final HashFileService hashService = hashServiceProvider.hashFileService();
         final String userPwd = create ?
                 hashService.createDirectories(userHash.getPwd(), userHash.getSalt(), userHash.saltResource(EnvironConstant.USER_CONF)).get(0) :
                 hashService.getHashPath(userHash.getPwd(), userHash.getMd5Salt(), userHash.saltResource(EnvironConstant.USER_CONF));
-
         final String userPropsPath = hashService.getHashPath(userPwd, userHash.getMd5Salt(), EnvironConstant.USER_PROP_FILENAME);
+        final CipherAES<String> cipher = hashServiceProvider.cipherService().createCipherResource(userHash.getSalt(), EnvironConstant.USER_PROP_FILENAME, userPropsPath, null);
+
         final File userPropFile = new File(userPropsPath);
-        final Properties userProps = new Properties();
+        Properties userProps = new Properties();
+
+        if (!create)
+            userProps = loadUserProperties(context, userPropsPath, cipher);
+
+        final String userHome = hashService.getHashPath(context.getUser().getPath(), userHash.getMd5Salt(), userHash.saltResource(EnvironConstant.USER_HOME));
+        final String userCert = hashService.getHashPath(userHome, userHash.getMd5Salt(), userHash.saltResource(EnvironConstant.USER_CERT));
 
         try {
-            final CipherAES<String> cipher = hashServiceProvider.cipherService().createCipherResource(userHash.getSalt(), EnvironConstant.USER_PROP_FILENAME, userPropsPath, null);
-
-            if (userPropFile.exists()) {
-                userProps.load(new ByteArrayInputStream(AESProvider.decryptFileBytes(Paths.get(userPropsPath), cipher.getCipherOut())));
-                context.setCertificate(hashServiceProvider.certificateService().getUserCertificate(context, userProps.getProperty(EnvironConstant.USER_CERT), userHash.getSalt()));
-            }
-            else {
+            if (create && !userPropFile.exists()) {
                 FileUtils.getInstance().createDirectory(context.getUser().getPath());
-                final String userHome = hashService.createDirectories(context.getUser().getPath(), userHash.getSalt(), userHash.saltResource(EnvironConstant.USER_HOME)).get(0);
-                final String userCert = hashService.createDirectories(userHome, userHash.getSalt(), userHash.saltResource(EnvironConstant.USER_CERT)).get(0);
-                final String identityPath = hashServiceProvider.identityService().createUserIdentity(context.getUser(), userHome, userHash.getSalt());
+               hashService.createDirectories(context.getUser().getPath(), userHash.getSalt(), userHash.saltResource(EnvironConstant.USER_HOME));
+               hashService.createDirectories(userHome, userHash.getSalt(), userHash.saltResource(EnvironConstant.USER_CERT));
+                final Identity identity = hashServiceProvider.identityService().createUserIdentity(context, userHome, userHash.getSalt());
                 final Certificate certificate = hashServiceProvider.certificateService().createUserCertificate(context, userCert, userHash.getSalt());
 
-                context.setCertificate(hashServiceProvider.certificateService().getUserCertificate(context, certificate, userHash.getSalt()));
+                context.setIdentity(hashServiceProvider.identityService().getUserIdentity(identity))
+                        .setCertificate(hashServiceProvider.certificateService().getUserCertificate(context, certificate));
 
                 userProps.setProperty(EnvironConstant.USER_HOME, userHome);
+                userProps.setProperty(EnvironConstant.USER_WORKSPACE, context.getUser().getPath());
                 userProps.setProperty(EnvironConstant.USER_CONF, userPwd);
-                userProps.setProperty(EnvironConstant.USER_IDENTITY_PROP, identityPath);
-                userProps.setProperty(EnvironConstant.USER_CERT, ((AbstractCipherResource<String>)certificate).getPath());
+                userProps.setProperty(EnvironConstant.USER_IDENTITY_PROP, ((AbstractCipherResource<String>) identity).getPath());
+                userProps.setProperty(EnvironConstant.USER_CERT, ((AbstractCipherResource<String>) certificate).getPath());
 
                 userProps.store(new FileWriter(userPropsPath), context.getUser().getEmail() + " - env properties");
                 AESProvider.encryptFile(userPropsPath, userPropsPath, cipher.getCipherIn());
@@ -96,8 +99,28 @@ public class EnvironServiceImpl implements EnvironService {
                 // hashServiceProvider.hashFileService().exists(user.getPath(), userHash.getMd5Salt(), userHash.saltResource(EnvironConstant.USER_HOME))
                 // hashService.exists(userHome, userHash.getMd5Salt(), userHash.saltResource(EnvironConstant.USER_CERT))
             }
+            else
+                context.setIdentity(hashServiceProvider.identityService().createUserIdentity(context, userHome, userHash.getSalt()))
+                    .setCertificate(hashServiceProvider.certificateService().createUserCertificate(context, userCert, userHash.getSalt()));
+
 
             return userProps;
+        }
+        catch (Exception e) {
+            LOG.error(e.getMessage());
+        }
+
+        return null;
+    }
+
+    private Properties loadUserProperties(final UserContext context, final String userPropsPath, final CipherAES<String> cipher) {
+        try {
+            final Properties props = new Properties();
+
+            props.load(new ByteArrayInputStream(AESProvider.decryptFileBytes(Paths.get(userPropsPath), cipher.getCipherOut())));
+            context.getUser().setPath(props.getProperty(EnvironConstant.USER_WORKSPACE));
+
+            return props;
         }
         catch (Exception e) {
             LOG.error(e.getMessage());
